+++ src/Stack/Ls.hs
@@ -21,8 +21,9 @@ import qualified Data.Text as T
 import qualified Data.Text.IO as T
 import Data.Typeable (Typeable)
 import qualified Data.Vector as V
+import Network.HTTP.StackClient (httpJSON)
 import Network.HTTP.Simple
-       (addRequestHeader, getResponseBody, httpJSON, parseRequest,
+       (addRequestHeader, getResponseBody, parseRequest,
         setRequestManager)
 import Network.HTTP.Types.Header (hAccept)
 import qualified Options.Applicative as OA